Rename example package to samples
